requirements.txt
fastapi[standard]==0.115.5
python-multipart==0.0.17
python-dotenv==1.0.1
